+++ js/app/svelte.config.js
@@ -1,5 +1,11 @@
-import adapter from "@sveltejs/adapter-auto";
+import adapter from "@sveltejs/adapter-node";
 import { vitePreprocess } from "@sveltejs/vite-plugin-svelte";
+
+import { resolve } from "path";
+import { fileURLToPath } from "url";
+const __dirname = fileURLToPath(import.meta.url);
+
+const out_path = resolve(__dirname, "../../../gradio/templates/node/build");
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
@@ -11,7 +17,9 @@ const config = {
 		// adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
 		// If your environment is not supported, or you settled on a specific environment, switch out the adapter.
 		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
-		adapter: adapter()
+		adapter: adapter({
+			out: out_path
+		})
 	}
 };